csgToBufferGeometry now uses an index in the result.
For a face with N sides,

With no index:  There are (N-2) triangles, each triangle uses 3
vertices, each vertex is 32 bytes (12 for pos, 12 for normal, 8 for uv).
So each triangle uses 3*32=96 bytes.  So that's (N-2)*96 bytes.

With an index:  There are N vertices of 32 bytes each.  Then there are
(N-2) sets of 3 index entries that are each 2 bytes.  So that's N*32 +
(N-2)*6 = N*36-12 bytes.

So we lose 6 bytes if a face is just a triangle, but we save roughly 60 bytes
per edge over 3.  These lumber CSGs are likely to have quad faces since
they start with them.